Update NMR-STAR Dicrionary (2018-02-27) and enabled differential update

+++ clone_bmrb.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+sync_update=true
+
 source ./scripts/db-user.sh
 
 PREFIX=bmr
@@ -85,7 +87,7 @@ fi
 XSD_SCHEMA=schema/mmcif_nmr-star.xsd
 DB_SCHEMA=schema/bmrb_clone.schema
 
-java -cp extlibs/xsd2pgschema.jar xsd2pgschema --xsd $XSD_SCHEMA --no-rel --hash-by SHA-1 --ddl $DB_SCHEMA
+java -cp extlibs/xsd2pgschema.jar xsd2pgschema --xsd $XSD_SCHEMA --no-rel --hash-by SHA-1 --ddl $DB_SCHEMA --inplace-doc-key-name entry_id --inplace-doc-key-name entry.id
 
 echo
 echo "Do you want to update $DB_NAME? (y [n]) "
@@ -100,28 +102,55 @@ esac
 
 ./$PREFIX"unzip_xml.sh" -a $ATOM
 
-psql -d $DB_NAME -U $DB_USER -f $DB_SCHEMA --quiet
+relations=`psql -d $DB_NAME -U $DB_USER -c "\d" | wc -l`
+
+if [ $sync_update != "true" ] || [ $relations = "0" ] ; then
+ sync_update=false
+ psql -d $DB_NAME -U $DB_USER -f $DB_SCHEMA --quiet
+fi
 
 WORK_DIR=pg_work
-CSV_DIR=$WORK_DIR/csv
+
+if [ $sync_update != "true" ] ; then
+ CSV_DIR=$WORK_DIR/csv
+fi
+
+MD5_DIR=chk_sum_pgsql_$PREFIX
 ERR_DIR=$WORK_DIR/err
 
 rm -rf $WORK_DIR
 
 mkdir -p $WORK_DIR
-mkdir -p $CSV_DIR
+
+if [ $sync_update != "true" ] ; then
+ mkdir -p $CSV_DIR
+fi
+
 mkdir -p $ERR_DIR
 
-rm -rf $CSV_DIR/*
+if [ $sync_update != "true" ] ; then
+ rm -rf $CSV_DIR/*
+fi
+
 rm -rf $ERR_DIR/*
 
 err_file=$ERR_DIR/all_err
 
-java -cp extlibs/xsd2pgschema.jar xml2pgcsv --xsd $XSD_SCHEMA --xml $XML_RAW_DIR --csv-dir $CSV_DIR --no-rel --no-valid --xml-file-ext-digest $FILE_EXT_DIGEST --db-name $DB_NAME --db-user $DB_USER 2> $err_file
+if [ $sync_update != "true" ] ; then
+
+ java -cp extlibs/xsd2pgschema.jar xml2pgcsv --xsd $XSD_SCHEMA --xml $XML_RAW_DIR --csv-dir $CSV_DIR --sync $MD5_DIR --no-rel --inplace-doc-key-name entry_id --inplace-doc-key-name entry.id --no-valid --xml-file-prefix-digest bmr --xml-file-ext-digest $FILE_EXT_DIGEST --db-name $DB_NAME --db-user $DB_USER 2> $err_file
+
+else
+
+ java -cp extlibs/xsd2pgschema.jar xml2pgsql --xsd $XSD_SCHEMA --xml $XML_RAW_DIR --sync $MD5_DIR --no-rel --inplace-doc-key-name entry_id --inplace-doc-key-name entry.id --no-valid --xml-file-prefix-digest bmr --xml-file-ext-digest $FILE_EXT_DIGEST --db-name $DB_NAME --db-user $DB_USER 2> $err_file
+
+fi
 
 if [ $? = 0 ] && [ ! -s $err_file ] ; then
  rm -f $err_file
- rm -rf $CSV_DIR
+ if [ $sync_update != "true" ] ; then
+  rm -rf $CSV_DIR
+ fi
 else
  echo "$0 aborted."
  exit 1
